+++ Source/SourceKittenFramework/Structure.swift
@@ -31,7 +31,7 @@ public struct Structure {
     - throws: Request.Error
     */
     public init(file: File) throws {
-        self.init(sourceKitResponse: try Request.editorOpen(file: file).failableSend())
+        self.init(sourceKitResponse: try Request.editorOpen(file: file).send())
     }
 }