+++ Source/sourcekitten/IndexCommand.swift
@@ -40,7 +40,7 @@ struct IndexCommand: CommandProtocol {
         let absoluteFile = options.file.bridge().absolutePathRepresentation()
         let request = Request.index(file: absoluteFile, arguments: options.compilerargs.components(separatedBy: " "))
         do {
-            print(toJSON(toNSDictionary(try request.failableSend())))
+            print(toJSON(toNSDictionary(try request.send())))
             return .success(())
         } catch {
             return .failure(.failed(error))