+++ Tests/SourceKittenFrameworkTests/DocInfoTests.swift
@@ -16,7 +16,7 @@ class DocInfoTests: XCTestCase {
     func testDocInfoRequest() {
         let swiftFile = File(path: fixturesDirectory + "DocInfo.swift")!
         let info = toNSDictionary(
-            try! Request.docInfo(text: swiftFile.contents, arguments: ["-sdk", sdkPath()]).failableSend()
+            try! Request.docInfo(text: swiftFile.contents, arguments: ["-sdk", sdkPath()]).send()
         )
         compareJSONString(withFixtureNamed: "DocInfo", jsonString: toJSON(info))
     }
@@ -29,7 +29,7 @@ class DocInfoTests: XCTestCase {
                                         "-c", swiftFile,
                                         "-module-name", "DocInfo",
                                         "-sdk", sdkPath()
-                                    ]).failableSend()
+                ]).send()
         )
         compareJSONString(withFixtureNamed: "ModuleInfo", jsonString: toJSON(info))
     }