+++ Tests/SourceKittenFrameworkTests/SourceKitTests.swift
@@ -184,7 +184,7 @@ class SourceKitTests: XCTestCase {
     func testIndex() {
         let file = "\(fixturesDirectory)Bicycle.swift"
         let arguments = ["-sdk", sdkPath(), "-j4", file ]
-        let indexJSON = NSMutableString(string: toJSON(toNSDictionary(try! Request.index(file: file, arguments: arguments).failableSend())) + "\n")
+        let indexJSON = NSMutableString(string: toJSON(toNSDictionary(try! Request.index(file: file, arguments: arguments).send())) + "\n")
 
         func replace(_ pattern: String, withTemplate template: String) {
             let regex = try! NSRegularExpression(pattern: pattern, options: [])
@@ -203,7 +203,7 @@ class SourceKitTests: XCTestCase {
     func testYamlRequest() {
         let path = fixturesDirectory + "Subscript.swift"
         let yaml = "key.request: source.request.editor.open\nkey.name: \"\(path)\"\nkey.sourcefile: \"\(path)\""
-        let output = try! Request.yamlRequest(yaml: yaml).failableSend()
+        let output = try! Request.yamlRequest(yaml: yaml).send()
         let expectedStructure = try! Structure(file: File(path: path)!)
         let actualStructure = Structure(sourceKitResponse: output)
         XCTAssertEqual(expectedStructure, actualStructure)